Fixes to heartbeats and NEB socket status logging
Switch to a 32-bit sequence number in heartbeats.

Future versions of ZeroMQ get rid of zmq_event_t, and we're already
doing the right thing, just wrapping it up in zmq_event_t. Get rid
of this weird opaque structure and future proof the socket status
logging code.

+++ mods/nagmq_pull.c
@@ -26,7 +26,7 @@ extern int errno;
 
 static void process_ping(json_t * payload) {
 	char * target;
-	int64_t seq;
+	int32_t seq;
 	char * extra = NULL;
 	struct timeval curtime;
